requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
httpx==0.25.2
jinja2==3.1.2
aiofiles==23.2.1
aiosqlite==0.21.0